// src/app/shared/models/part.model.ts
export interface PartModel {
  descripcion: string;
  precio_sugerido: number;
  ultima_compra?: string;
  ImpuestoInterno: string;
  StockMinimo: number;
  ConceptoCompra: number;
  costo_campania: number;
  PublicadoMercadoLibre?: number;
  reemplazo_a_id?: string;
  marca?: string;
  ImporteTotalCompra: number;
  descuento_porcentaje?: number;
  precio_oferta: number;
  StockMaximo: number;
  codigo: string;
  CostoFOB: number;
  precio_lista: number;
  MonedaCompra: number;
  repuesto_id: number;
  descuento_adicional: number;
  itemId?: string;
  observaciones: string;
  ultima_venta?: string;
  Importado: boolean;
  fecha_lista?: string;
  vigente: number;
  origen?: string;
  statusMercadoLibre?: string;
  ultimo_costo: number;
  ConceptoVenta: number;
  tipo_articulo: string;
  clase: string;
  DespachoNro: string;
  moneda: number;
  CantidadTotalCompra: number;
  dpm: number;
  Reemplazoa: string;
  original: number;
  CoeficienteNacionalizacion: number;
  descuento?: string;
  tasa_iva_id: number;
  control_inventario: string;
  precio_alternativo: number;
  ultimo_precio: number;
  costo_oferta: number;
  reemplazado_por_id?: string;
  check_dpm: number;
  Reemplazadopor: string;
  origen_descripcion?: string;
  ConceptoCostoMostrador: number;
  familia?: string;
  quantity: number;
  localidad: string;
}